src/main/webapp/app/entities/redeem/delete/redeem-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { RedeemService } from '../service/redeem.service';

import { RedeemDeleteDialogComponent } from './redeem-delete-dialog.component';

describe('Redeem Management Delete Component', () => {
  let comp: RedeemDeleteDialogComponent;
  let fixture: ComponentFixture<RedeemDeleteDialogComponent>;
  let service: RedeemService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RedeemDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(RedeemDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(RedeemDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(RedeemService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete('ABC');
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith('ABC');
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
